// src/types/flower.ts
export type FlowerData = {
  id: string
  x: number
  y: number
  rotation: number
  scale: number
}

export type HeartAnimationFlower = {
  id: string
  scale: number
  rotation: number
  burstRotation: number
  heartRotation: number
  startX: number
  startY: number
  burstX: number
  burstY: number
  heartX: number
  heartY: number
  appearDelay: number
  burstDuration: number
  formDelay: number
  formDuration: number
}